+++ src/babylon_js/materials/nodes/glossy.py
@@ -1,4 +1,4 @@
-from .abstract  import AbstractBJSNode, SPECULAR_TEX
+from .abstract import AbstractBJSNode, ROUGHNESS_TEX, SPECULAR_TEX
 
 from mathutils import Color
 
@@ -13,6 +13,12 @@ class GlossyBJSNode(AbstractBJSNode):
         defaultColor = self.findTexture(input, SPECULAR_TEX)
         if defaultColor is not None:
             self.specularColor = Color((defaultColor[0], defaultColor[1], defaultColor[2]))
-
-        self.roughness = self.findInput('Roughness')
+            
         self.mustBakeSpecular = input.mustBake if isinstance(input, AbstractBJSNode) else False
+        # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+        input = self.findInput('Roughness')
+        defaultRoughness = self.findTexture(input, ROUGHNESS_TEX)
+        if defaultRoughness is not None:
+            self.roughness = defaultRoughness
+
+        self.mustBakeRoughness = input.mustBake if isinstance(input, AbstractBJSNode) else False